debug: fix all trobbles encontred

+++ internal/domain/Permission.go
@@ -1,6 +1,6 @@
 package domain
 
-var Permissions = map[string]byte{
+var Permissions = map[string]uint64{
 	"user":  1 << 0,
 	"mod":   1 << 1,
 	"dev":   1 << 2,
